requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.7
drf-spectacular==0.26.4
python-dotenv==1.0.0
Pillow==10.0.1
setuptools==68.2.2